requirements.txt
ib_insync>=0.9.86
scipy>=1.10
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
streamlit>=1.30
streamlit-autorefresh>=1.0
plotly>=5.18
